requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
frozenlist==1.4.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
multidict==6.0.5
nltk==3.9.1
numpy==2.1.0
pandas==2.2.2
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.7.24
requests==2.32.3
six==1.16.0
tqdm==4.66.5
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
yarl==1.9.6
